+++ inst/doc/getquandldata-vignette-introduction.R
@@ -1,16 +1,15 @@
-## ---- echo = FALSE-------------------------------------------------------
+## ---- echo = FALSE------------------------------------------------------------
 knitr::opts_chunk$set(message = FALSE, cache = FALSE,eval=FALSE)
 
-## ---- eval=FALSE---------------------------------------------------------
+## ---- eval=FALSE--------------------------------------------------------------
 #  # not in CRAN yet (need to test it further)
 #  #install.packages('GetQuandlData')
 #  
 #  # from github
 #  devtools::install_github('msperlin/GetQuandlData')
 
-## ---- eval=FALSE---------------------------------------------------------
+## ---- eval=FALSE--------------------------------------------------------------
 #  library(GetQuandlData)
-#  library(tidyverse)
 #  
 #  my_id <- c('Inflation USA' = 'RATEINF/INFLATION_USA')
 #  my_api <- readLines('YOURAPIHERE') # you need your own API (get it at https://www.quandl.com/sign-up-modal?defaultModal=showSignUp>)
@@ -23,9 +22,9 @@ knitr::opts_chunk$set(message = FALSE, cache = FALSE,eval=FALSE)
 #                          last_date = last_date,
 #                          cache_folder = tempdir())
 #  
-#  glimpse(df)
+#  dplyr::glimpse(df)
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  p <- ggplot(df, aes(x = ref_date, y = value/100)) +
 #    geom_col() +
 #    labs(y = 'Inflation (%)',
@@ -35,9 +34,8 @@ knitr::opts_chunk$set(message = FALSE, cache = FALSE,eval=FALSE)
 #  
 #  p
 
-## ---- message=TRUE, eval=FALSE-------------------------------------------
+## ---- message=TRUE, eval=FALSE------------------------------------------------
 #  library(GetQuandlData)
-#  library(tidyverse)
 #  
 #  db_id <- 'RATEINF'
 #  my_api <- readLines('YOURAPIHERE') # you need your own API
@@ -46,12 +44,12 @@ knitr::opts_chunk$set(message = FALSE, cache = FALSE,eval=FALSE)
 #  
 #  head(df)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 #  idx <- stringr::str_detect(df$name, 'Inflation YOY')
 #  
 #  df_series <- df[idx, ]
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 #  my_id <- df_series$quandl_code
 #  names(my_id) <- df_series$name
 #  first_date <- '2010-01-01'
@@ -65,7 +63,7 @@ knitr::opts_chunk$set(message = FALSE, cache = FALSE,eval=FALSE)
 #  
 #  glimpse(df_inflation)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 #  p <- ggplot(df_inflation, aes(x = ref_date, y = value/100)) +
 #    geom_col() +
 #    labs(y = 'Inflation (%)',