+++ NEWS.md
@@ -1,3 +1,10 @@
+## Version 1.0.0 (2023-02-15)
+
+- fixed change in api call
+- removed date parameters
+- removed dependency of tidyverse on DESCRIPTION
+- added tests and GH actions
+
 ## Version 0.1.0 (2019-10-13)
 
 - First commit